+++ src/lib/config.ts
@@ -60,8 +60,8 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 export const SDK_METADATA = {
   language: "typescript",
   openapiDocVersion: "1.0.0",
-  sdkVersion: "0.0.1-alpha.22",
-  genVersion: "2.882.0",
+  sdkVersion: "0.0.1-alpha.23",
+  genVersion: "2.884.4",
   userAgent:
-    "speakeasy-sdk/typescript 0.0.1-alpha.22 2.882.0 1.0.0 @gbg/go-core",
+    "speakeasy-sdk/typescript 0.0.1-alpha.23 2.884.4 1.0.0 @gbg/go-core",
 } as const;